Flatten_Size/requirements.txt
streamlit>=1.32.0
streamlit-drawable-canvas>=0.9.2
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0